2022.03.24. - Add components for cell content (list, tags) - Modify component for pi table
