+++ src/views/TestPI.tsx
@@ -1,6 +1,6 @@
 // Component
 import Page from '../components/common/Page';
-import { PersonalInfoTable } from '../components/common/Table';
+import { PersonalInfoTable } from '../components/PITable';
 
 const TestPI = (): JSX.Element => {
   return (
